requirements.txt
attrs==19.3.0
django==3.0.3
